Coding Days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2662 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/finance/application/BlackScholesSolverWithStretching.cpp
@@ -562,7 +562,7 @@ void BlackScholesSolverWithStretching::initLogTransformedGridWithPayoff(sg::base
 	}
 }
 
-void BlackScholesSolverWithStretching::printGrid(sg::base::DataVector& alpha, double PointesPerDimension, std::string tfilename) const
+void BlackScholesSolverWithStretching::printGrid(sg::base::DataVector& alpha, size_t PointesPerDimension, std::string tfilename) const
 {
 	sg::base::GridPrinterForStretching myPrinter(*this->myGrid);
 	myPrinter.printGrid(alpha, tfilename, PointesPerDimension);
